requirements.txt
streamlit
pillow
folium
streamlit-folium
piexif
